Added before_request, g objects, more detailed navbar, removed last changed in files
Removed Code repetition in routes which are in before_request already. Also the try-except blocks "save islands" were removed, because a 500-errorhandler was implemented.
Additionally within the before_request the task of check jwt and 2fa were added and saved in the g object. So it will be easier inside the routes so handle it.
With the g objects, it was easier to implement a more precise navbar deciding what to show based on what you are authenticated with.